# Optimize performance of `lineAndCharacterForCharacterOffset(_:)`