Rename `ByteOffsetCache` to `CacheContainer`
Add `lineAndCharacterForCharacterOffset(_:)` to `CacheContainer`

Linting KeychainAccess by SwiftLint(3837c73), SourceKitten(668d974)
Before applying this commit:
```
$ time swiftlint
…
swiftlint  10.74s user 0.17s system 95% cpu 11.438 total
```

After this commit applied:
```
$ time swiftlint
…
swiftlint  7.89s user 0.17s system 95% cpu 8.453 total
```